+++ console/pom.xml
@@ -14,7 +14,6 @@
 	<modules>
 		<module>backend</module>
 		<module>frontend</module>
-		<module>ladybug</module>
 		<module>war</module>
 	</modules>
