+++ ladybug/common/src/main/java/org/frankframework/ladybug/config/LadybugConfiguration.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package org.frankframework.ladybug;
+package org.frankframework.ladybug.config;
 
 import lombok.extern.log4j.Log4j2;
 import nl.nn.testtool.web.ApiServlet;
