+++ ladybug/common/src/main/java/org/frankframework/ladybug/config/LadybugSecurityChainConfigurer.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2024 WeAreFrank!
+   Copyright 2023-2024 WeAreFrank!
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package org.frankframework.ladybug;
+package org.frankframework.ladybug.config;
 
 import java.lang.reflect.Method;
 import java.util.stream.Collectors;
